Cleanup part 1, UI is still completely broken
