+++ RandomizerCore/Controllers/ControllerBase.cs
@@ -1,0 +1,429 @@
+﻿using System.Reflection;
+using RandomizerCore.Core;
+using RandomizerCore.Randomizer;
+using RandomizerCore.Randomizer.Models;
+using RandomizerCore.Randomizer.Shuffler;
+using RandomizerCore.Utilities.Logging;
+using RandomizerCore.Utilities.Models;
+using RandomizerCore.Utilities.Util;
+
+namespace RandomizerCore.Controllers.Models;
+
+public abstract class ControllerBase
+{
+    #if DEBUG
+        public string AppName => "Minish Cap Randomizer Debug Build";
+    #else
+	    public string AppName => "Minish Cap Randomizer";
+    #endif
+
+    public string VersionName => VersionIdentifier;
+    public string RevName => RevisionIdentifier;
+
+    public static string VersionIdentifier => "v0.7.0";
+    public static string RevisionIdentifier => "Pre-release";
+
+    public void SetLogOutputPath(string logFilePath)
+    {
+        Logger.Instance.OutputFilePath = logFilePath;
+    }
+
+    public void SetLoggerVerbosity(bool useVerboseLogger)
+    {
+        Logger.Instance.UseVerboseLogger = useVerboseLogger;
+    }
+
+    public ShufflerControllerResult LoadSettingsFromSettingString(string settingString)
+    {
+        try
+        {
+            MinifiedSettings.GenerateSettingsFromBase64String(settingString, Shuffler.GetSelectedOptions().OnlyLogic().GetSorted(),
+                Shuffler.GetSelectedOptions().OnlyLogic().GetCrc32());
+
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            Logger.Instance.SaveLogTransaction();
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+    }
+    
+    public ShufflerControllerResult LoadCosmeticsFromCosmeticsString(string settingString)
+    {
+        try
+        {
+            MinifiedSettings.GenerateSettingsFromBase64String(settingString, Shuffler.GetSelectedOptions().OnlyCosmetic().GetSorted(),
+                Shuffler.GetSelectedOptions().OnlyCosmetic().GetCrc32());
+
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            Logger.Instance.SaveLogTransaction();
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+    }    
+    
+    public string GetSelectedSettingsString()
+    {
+        return MinifiedSettings.GenerateSettingsString(Shuffler.GetSelectedOptions().OnlyLogic().GetSorted(),
+            Shuffler.GetSelectedOptions().OnlyLogic().GetCrc32());
+    }
+
+    public string GetSelectedCosmeticsString()
+    {
+        return MinifiedSettings.GenerateSettingsString(Shuffler.GetSelectedOptions().OnlyCosmetic().GetSorted(),
+            Shuffler.GetSelectedOptions().OnlyCosmetic().GetCrc32());
+    }
+
+    public string GetFinalSettingsString()
+    {
+        return MinifiedSettings.GenerateSettingsString(Shuffler.GetFinalOptions().OnlyLogic().GetSorted(),
+            Shuffler.GetFinalOptions().OnlyLogic().GetCrc32());
+    }
+
+    public string GetFinalCosmeticsString()
+    {
+        return MinifiedSettings.GenerateSettingsString(Shuffler.GetFinalOptions().OnlyCosmetic().GetSorted(),
+            Shuffler.GetFinalOptions().OnlyCosmetic().GetCrc32());
+    }
+
+    public ShufflerControllerResult LoadRom(string filename)
+    {
+        try
+        {
+            Rom.Initialize(filename);
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+    public uint GetSettingsChecksum()
+    {
+        return Shuffler.GetLogicOptionsCrc32();
+    }
+
+    public uint GetCosmeticsChecksum()
+    {
+        return Shuffler.GetCosmeticOptionsCrc32();
+    }
+
+    public void FlushLogger()
+    {
+        Logger.Instance.Flush();
+    }
+
+    public void SetRandomizationSeed(ulong seed)
+    {
+        Shuffler.SetSeed(seed);
+    }
+
+    public OptionList GetSelectedOptions()
+    {
+        return Shuffler.GetSelectedOptions();
+    }
+
+    public OptionList GetFinalOptions()
+    {
+        return Shuffler.GetFinalOptions();
+    }
+
+    public ShufflerControllerResult LoadLogicFile(string? filename = null)
+    {
+        try
+        {
+            Shuffler.LoadOptions(filename);
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+    public ShufflerControllerResult SaveAndPatchRom(string filename, string? patchFile = null)
+    {
+        try
+        {
+            Shuffler.ValidateState(true);
+            var romBytes = Shuffler.GetRandomizedRom();
+            File.WriteAllBytes(filename, romBytes);
+            int exitCode = Shuffler.ApplyPatch(filename, patchFile);
+            if (exitCode != 0)
+                throw new Exception("Errors occured when saving the rom");
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+    public ShufflerControllerResult CreatePatch(string patchFilename, string? patchFile = null)
+    {
+        try
+        {
+            Shuffler.ValidateState(true);
+            var romBytes = Shuffler.GetRandomizedRom();
+            var stream = new MemoryStream(romBytes);
+            int exitCode = Shuffler.ApplyPatch(stream, patchFile);
+            if (exitCode != 0)
+                throw new Exception("Errors occured when saving the rom");
+            var patch = BpsPatcher.GeneratePatch(Rom.Instance!.RomData, romBytes, patchFilename);
+            File.WriteAllBytes(patchFilename, patch.Content);
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+    public PatchFile CreatePatch()
+    {
+        Shuffler.ValidateState(true);
+        var romBytes = Shuffler.GetRandomizedRom();
+        var stream = new MemoryStream(romBytes);
+        Shuffler.ApplyPatch(stream);
+        return BpsPatcher.GeneratePatch(Rom.Instance!.RomData, romBytes, "Patch");
+    }
+
+    public string CreateSpoiler()
+    {
+        Shuffler.ValidateState(true);
+        return Shuffler.GetSpoiler();
+    }
+
+    /// <summary>
+    ///     Creates a patch from a patched ROM
+    /// </summary>
+    /// <param name="patchFilename"></param>
+    /// <param name="patchedRomFilename"></param>
+    /// <returns></returns>
+    public ShufflerControllerResult SaveRomPatch(string patchFilename, string patchedRomFilename)
+    {
+        if (Rom.Instance == null)
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                ErrorMessage =
+                    "Cannot patch ROM! No base ROM was loaded! Please select a vanilla European Minish Cap ROM and try again."
+            };
+
+        try
+        {
+            var patchedRom = File.ReadAllBytes(patchedRomFilename);
+            var patch = BpsPatcher.GeneratePatch(Rom.Instance.RomData, patchedRom, patchFilename);
+            File.WriteAllBytes(patchFilename, patch.Content);
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+    /// <summary>
+    ///     Applies a BPS patch to a vanilla EU Minish Cap ROM
+    /// </summary>
+    /// <param name="outputFilename"></param>
+    /// <param name="patchFilename"></param>
+    /// <returns></returns>
+    public ShufflerControllerResult PatchRom(string outputFilename, string patchFilename)
+    {
+        if (Rom.Instance == null)
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                ErrorMessage =
+                    "Cannot patch ROM! No base ROM was loaded! Please select a vanilla European Minish Cap ROM and try again."
+            };
+
+        try
+        {
+            var patchContent = File.ReadAllBytes(patchFilename);
+            var patchedRom = BpsPatcher.ApplyPatch(Rom.Instance.RomData, new PatchFile { Content = patchContent });
+            File.WriteAllBytes(outputFilename, patchedRom);
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+    public ShufflerControllerResult SaveSpoiler(string filename)
+    {
+        try
+        {
+            Shuffler.ValidateState(true);
+            var spoiler = Shuffler.GetSpoiler();
+            File.WriteAllText(filename, spoiler);
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+    public ShufflerControllerResult ExportDefaultLogic(string filepath)
+    {
+        try
+        {
+            var assembly = Assembly.GetAssembly(typeof(ControllerBase));
+            using var stream = assembly?.GetManifestResourceStream("RandomizerCore.Resources.default.logic");
+            File.WriteAllText(filepath, new StreamReader(stream).ReadToEnd());
+            return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+            return new ShufflerControllerResult
+            {
+                WasSuccessful = false,
+                Error = e,
+                ErrorMessage = e.Message
+            };
+        }
+        finally
+        {
+            Logger.Instance.SaveLogTransaction();
+        }
+    }
+
+	public ShufflerControllerResult ExportYAML(string filepath, bool mystery)
+	{
+        string? name = null;
+        var index = filepath.LastIndexOf(Path.DirectorySeparatorChar);
+        if (index != -1)
+        {
+            name = filepath.Substring(index + 1);
+            index = name.LastIndexOf(".");
+            if(index != -1)
+                name = name.Substring(0, index);
+        }
+        try
+        {
+			File.WriteAllText(filepath, Mystery.CreateYAML(name, null, GetSelectedOptions(), mystery));
+			return new ShufflerControllerResult { WasSuccessful = true };
+        }
+        catch (Exception e)
+        {
+            Logger.Instance.LogException(e);
+			return new ShufflerControllerResult
+			{
+				WasSuccessful = false,
+				Error = e,
+                ErrorMessage = e.Message,
+			};
+        }
+		finally
+		{
+			Logger.Instance.SaveLogTransaction();
+		}
+    }
+
+    protected ControllerBase(Type baseClassType)
+    {
+        Logger.Instance.LogInfo($"Minish Cap Randomizer Core Version {VersionName} {RevName} initialized!");
+        Logger.Instance.LogInfo($"Shuffler {baseClassType} initialized!");
+        Logger.Instance.SaveLogTransaction(true);
+    }
+    
+    internal ShufflerBase Shuffler { get; set; }
+
+    public abstract ulong FinalSeed { get; }
+
+    public abstract string SeedFilename { get; }
+
+    public abstract ShufflerControllerResult LoadLocations(string? logicFile = null, string? yamlFileLogic = null, string? yamlFileCosmetics = null, bool useGlobalYAML = false);
+
+    public abstract ShufflerControllerResult Randomize(int retries = 1, bool useSphereBasedShuffler = false);
+}
